+++ CHANGELOG.md
@@ -3,6 +3,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [0.1.4] - 2019-05-29
+
+### Added
+- You can now configure which environment would get what externs.
+You can keep using `:externs-paths ["a" "b"]` which would set externs both
+for `:development` and `:production` builds. If you want to distinguish them,
+you can now do `:externs-paths {:development ["a" "b"] :production ["x" "y"]}`.
+
+### Changed
+- `:duct.server/figwheel` key is modified only when in `:development` environment.
+- `:duct.compiler/cljs` key is modified only when in `:production` environment.
+
 ## [0.1.3] - 2019-05-27
 
 ### Added
@@ -15,3 +27,4 @@ extern file in location `src/<project-dirs>/client/externs.js`.
 To retain it, you have to add `:add-example-api? true` to the config options.
 
 [0.1.3]: https://github.com/magnetcoop/hydrogen.module.cljs/releases/tag/v0.1.3
+[0.1.4]: https://github.com/magnetcoop/hydrogen.module.cljs/releases/tag/v0.1.4